requirements.txt
numpy
pandas
matplotlib
zarr
dask
fsspec
gcsfs
jupyterlab
jupyter-book
gsutil
scikit-allel
malariagen-data
